// open_contabilidad/src/interfaces/IRut.ts
export interface IRut{
    codigo: number;
    dv:string;
    nombre: string;
    direccion: string;
    ciudad: string;
    comuna: string;
    giro: string;
    telefonos: string;
    fax: string;
    email:string;
    tipo: number;
    banco:number;
    tipocuenta:number;
    nrocuenta:number;
    emailintercambio:string;
    condiciones:string;
}